FoundationEssentials: provide typed overloads for constants from WinSDK
This allows for us to use the constants without explicitly casting the
type to `DWORD` on each site of use. By providing the shadowing
overload we can simply use the constants without impacting the
readability of the surrounding code.

+++ Sources/FoundationEssentials/Error/CocoaError+FilePath.swift
@@ -140,39 +140,39 @@ extension CocoaError {
 
     static func errorWithFilePath(_ path: String? = nil, win32 dwError: DWORD, reading: Bool, variant: String? = nil, userInfo: [String : AnyHashable] = [:]) -> CocoaError {
         let code: CocoaError.Code = switch (reading, dwError) {
-            case (true, DWORD(ERROR_FILE_NOT_FOUND)), (true, DWORD(ERROR_PATH_NOT_FOUND)):
+            case (true, ERROR_FILE_NOT_FOUND), (true, ERROR_PATH_NOT_FOUND):
                 // Windows will return ERROR_FILE_NOT_FOUND or ERROR_PATH_NOT_FOUND
                 // for empty paths.
                 (path?.isEmpty ?? false) ? .fileReadInvalidFileName : .fileReadNoSuchFile
-            case (true, DWORD(ERROR_ACCESS_DENIED)): .fileReadNoPermission
-            case (true, DWORD(ERROR_INVALID_ACCESS)): .fileReadNoPermission
-            case (true, DWORD(ERROR_INVALID_DRIVE)): .fileReadNoSuchFile
-            case (true, DWORD(ERROR_SHARING_VIOLATION)): .fileReadNoPermission
-            case (true, DWORD(ERROR_INVALID_NAME)): .fileReadInvalidFileName
-            case (true, DWORD(ERROR_LABEL_TOO_LONG)): .fileReadInvalidFileName
-            case (true, DWORD(ERROR_BAD_PATHNAME)): .fileReadInvalidFileName
-            case (true, DWORD(ERROR_FILENAME_EXCED_RANGE)): .fileReadInvalidFileName
-            case (true, DWORD(ERROR_DIRECTORY)): .fileReadInvalidFileName
+            case (true, ERROR_ACCESS_DENIED): .fileReadNoPermission
+            case (true, ERROR_INVALID_ACCESS): .fileReadNoPermission
+            case (true, ERROR_INVALID_DRIVE): .fileReadNoSuchFile
+            case (true, ERROR_SHARING_VIOLATION): .fileReadNoPermission
+            case (true, ERROR_INVALID_NAME): .fileReadInvalidFileName
+            case (true, ERROR_LABEL_TOO_LONG): .fileReadInvalidFileName
+            case (true, ERROR_BAD_PATHNAME): .fileReadInvalidFileName
+            case (true, ERROR_FILENAME_EXCED_RANGE): .fileReadInvalidFileName
+            case (true, ERROR_DIRECTORY): .fileReadInvalidFileName
             case (true, _): .fileReadUnknown
 
-            case (false, DWORD(ERROR_FILE_NOT_FOUND)), (false, DWORD(ERROR_PATH_NOT_FOUND)):
+            case (false, ERROR_FILE_NOT_FOUND), (false, ERROR_PATH_NOT_FOUND):
                 // Windows will return ERROR_FILE_NOT_FOUND or ERROR_PATH_NOT_FOUND
                 // for empty paths.
                 (path?.isEmpty ?? false) ? .fileWriteInvalidFileName : .fileNoSuchFile
-            case (false, DWORD(ERROR_ACCESS_DENIED)): .fileWriteNoPermission
-            case (false, DWORD(ERROR_INVALID_ACCESS)): .fileWriteNoPermission
-            case (false, DWORD(ERROR_INVALID_DRIVE)): .fileNoSuchFile
-            case (false, DWORD(ERROR_WRITE_FAULT)): .fileWriteVolumeReadOnly
-            case (false, DWORD(ERROR_SHARING_VIOLATION)): .fileWriteNoPermission
-            case (false, DWORD(ERROR_FILE_EXISTS)): .fileWriteFileExists
-            case (false, DWORD(ERROR_DISK_FULL)): .fileWriteOutOfSpace
-            case (false, DWORD(ERROR_INVALID_NAME)): .fileWriteInvalidFileName
-            case (false, DWORD(ERROR_LABEL_TOO_LONG)): .fileWriteInvalidFileName
-            case (false, DWORD(ERROR_BAD_PATHNAME)): .fileWriteInvalidFileName
-            case (false, DWORD(ERROR_ALREADY_EXISTS)): .fileWriteFileExists
-            case (false, DWORD(ERROR_FILENAME_EXCED_RANGE)): .fileWriteInvalidFileName
-            case (false, DWORD(ERROR_DIRECTORY)): .fileWriteInvalidFileName
-            case (false, DWORD(ERROR_DISK_RESOURCES_EXHAUSTED)): .fileWriteOutOfSpace
+            case (false, ERROR_ACCESS_DENIED): .fileWriteNoPermission
+            case (false, ERROR_INVALID_ACCESS): .fileWriteNoPermission
+            case (false, ERROR_INVALID_DRIVE): .fileNoSuchFile
+            case (false, ERROR_WRITE_FAULT): .fileWriteVolumeReadOnly
+            case (false, ERROR_SHARING_VIOLATION): .fileWriteNoPermission
+            case (false, ERROR_FILE_EXISTS): .fileWriteFileExists
+            case (false, ERROR_DISK_FULL): .fileWriteOutOfSpace
+            case (false, ERROR_INVALID_NAME): .fileWriteInvalidFileName
+            case (false, ERROR_LABEL_TOO_LONG): .fileWriteInvalidFileName
+            case (false, ERROR_BAD_PATHNAME): .fileWriteInvalidFileName
+            case (false, ERROR_ALREADY_EXISTS): .fileWriteFileExists
+            case (false, ERROR_FILENAME_EXCED_RANGE): .fileWriteInvalidFileName
+            case (false, ERROR_DIRECTORY): .fileWriteInvalidFileName
+            case (false, ERROR_DISK_RESOURCES_EXHAUSTED): .fileWriteOutOfSpace
             case (false, _): .fileWriteUnknown
         }
